flask/requirements.txt
flask
flask-sqlalchemy
python-dotenv
# zappa # for AWS Lambda Flask apps